refactor: move buildPeriodData into usage-aggregator module
