+++ src/usage-aggregator.ts
@@ -1,0 +1,42 @@
+import { CATEGORY_LABELS, type ProjectSummary, type TaskCategory } from './types.js'
+import { type PeriodData } from './menubar-json.js'
+
+export function buildPeriodData(label: string, projects: ProjectSummary[]): PeriodData {
+  const sessions = projects.flatMap(p => p.sessions)
+  const catTotals: Record<string, { turns: number; cost: number; editTurns: number; oneShotTurns: number }> = {}
+  const modelTotals: Record<string, { calls: number; cost: number }> = {}
+  let inputTokens = 0, outputTokens = 0, cacheReadTokens = 0, cacheWriteTokens = 0
+
+  for (const sess of sessions) {
+    inputTokens += sess.totalInputTokens
+    outputTokens += sess.totalOutputTokens
+    cacheReadTokens += sess.totalCacheReadTokens
+    cacheWriteTokens += sess.totalCacheWriteTokens
+    for (const [cat, d] of Object.entries(sess.categoryBreakdown)) {
+      if (!catTotals[cat]) catTotals[cat] = { turns: 0, cost: 0, editTurns: 0, oneShotTurns: 0 }
+      catTotals[cat].turns += d.turns
+      catTotals[cat].cost += d.costUSD
+      catTotals[cat].editTurns += d.editTurns
+      catTotals[cat].oneShotTurns += d.oneShotTurns
+    }
+    for (const [model, d] of Object.entries(sess.modelBreakdown)) {
+      if (!modelTotals[model]) modelTotals[model] = { calls: 0, cost: 0 }
+      modelTotals[model].calls += d.calls
+      modelTotals[model].cost += d.costUSD
+    }
+  }
+
+  return {
+    label,
+    cost: projects.reduce((s, p) => s + p.totalCostUSD, 0),
+    calls: projects.reduce((s, p) => s + p.totalApiCalls, 0),
+    sessions: projects.reduce((s, p) => s + p.sessions.length, 0),
+    inputTokens, outputTokens, cacheReadTokens, cacheWriteTokens,
+    categories: Object.entries(catTotals)
+      .sort(([, a], [, b]) => b.cost - a.cost)
+      .map(([cat, d]) => ({ name: CATEGORY_LABELS[cat as TaskCategory] ?? cat, ...d })),
+    models: Object.entries(modelTotals)
+      .sort(([, a], [, b]) => b.cost - a.cost)
+      .map(([name, d]) => ({ name, ...d })),
+  }
+}
